+++ sveltekit/src/lib/helpers/constants.ts
@@ -70,6 +70,5 @@ export enum AccessKeySettings {
 }
 
 export const ACCESS_KEY_COOKIE_NAME = '_canutin-accessKey=';
-export const ACCESS_KEY_UNAUTHORIZED = 'Unauthorized';
 
 export const UNDOABLE_ACTION = "🚩 This action can't be undone\n";
